tests/app/test_app.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Android App 测试示例用例（基于 Appium）

注意：运行这些测试需要：
1. 安装 Appium Server
2. 配置 Android SDK
3. 连接真机或启动模拟器
4. 安装被测 App
"""

import pytest
from appium import webdriver
from appium.options.android import UiAutomator2Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import time


# Appium 配置（示例，需要根据实际环境修改）
DEFAULT_CAPS = {
    'platformName': 'Android',
    'platformVersion': '11',
    'deviceName': 'Android Emulator',
    'appPackage': 'com.example.app',  # 替换为实际包名
    'appActivity': '.MainActivity',   # 替换为实际 Activity
    'automationName': 'UiAutomator2',
    'noReset': True,
}


@pytest.fixture(scope="function")
def driver():
    """Appium Driver _fixture"""
    # 实际使用时需要配置正确的 Appium Server 地址
    appium_server = "http://localhost:4723/wd/hub"
    
    try:
        options = UiAutomator2Options().load_capabilities(DEFAULT_CAPS)
        driver = webdriver.Remote(appium_server, options=options)
        driver.implicitly_wait(10)
        yield driver
    except Exception as e:
        pytest.skip(f"无法连接到 Appium Server: {e}")
    finally:
        try:
            driver.quit()
        except:
            pass


@pytest.mark.skip(reason="需要真实设备/模拟器环境")
class TestApp:
    """Android App 测试类"""
    
    def test_app_launch(self, driver):
        """测试 App 启动"""
        # 等待主页面加载
        wait = WebDriverWait(driver, 10)
        # 这里需要根据实际 App 的元素修改
        # element = wait.until(EC.presence_of_element_located((By.ID, "com.example.app:id/home_view")))
        # assert element.is_displayed()
        assert True  # 示例占位
    
    def test_login_function(self, driver):
        """测试登录功能"""
        # 示例测试逻辑
        # 1. 找到用户名输入框
        # 2. 输入用户名
        # 3. 找到密码输入框
        # 4. 输入密码
        # 5. 点击登录按钮
        # 6. 验证登录成功
        
        # 占位实现
        assert True
    
    def test_navigation(self, driver):
        """测试页面导航"""
        # 测试 App 内页面跳转是否正常
        assert True
    
    def test_search_function(self, driver):
        """测试搜索功能"""
        # 1. 点击搜索框
        # 2. 输入搜索关键词
        # 3. 验证搜索结果
        
        assert True
    
    @pytest.mark.parametrize("test_data", [
        {"username": "test_user1", "password": "pass123"},
        {"username": "test_user2", "password": "pass456"},
    ])
    def test_login_with_multiple_users(self, driver, test_data):
        """参数化测试多用户登录"""
        # 使用不同用户数据测试登录
        assert True


@pytest.mark.skip(reason="性能测试需要特殊配置")
class TestAppPerformance:
    """App 性能测试"""
    
    def test_launch_time(self, driver):
        """测试 App 启动时间"""
        start_time = time.time()
        
        # 重启 App 并测量启动时间
        # driver.launch_app()
        # wait.until(...)
        
        elapsed = time.time() - start_time
        assert elapsed < 5.0, f"App 启动时间过长：{elapsed:.2f}秒"
    
    def test_memory_usage(self, driver):
        """测试内存使用"""
        # 获取 App 内存使用情况
        # performance_data = driver.get_performance_data('com.example.app', 'memoryinfo', 1)
        pass
    
    def test_cpu_usage(self, driver):
        """测试 CPU 使用率"""
        # 获取 CPU 使用情况
        pass


@pytest.mark.skip(reason="需要特殊配置")
class TestAppGestures:
    """手势操作测试"""
    
    def test_swipe(self, driver):
        """测试滑动操作"""
        # 实现滑动测试
        pass
    
    def test_tap(self, driver):
        """测试点击操作"""
        pass
    
    def test_long_press(self, driver):
        """测试长按操作"""
        pass
